+++ packages/patrol_finders/CHANGELOG.md
@@ -1,3 +1,14 @@
+## Unreleased
+
+- Fix issues with `enterText()`. (#2202, #2111)
+- Add `hideKeyboard` optional argument to `enterText()` method.
+
+## 3.4.0
+
+- Fix `at()`, `first` and `last` to wait for widget to become visible.
+- Add `isAndroid`, `isIOS`, `isWeb` and `isMacOS` getters.
+- Remove `dart:io`.
+
 ## 3.3.0
 
 - Bump `patrol_log` to `^0.9.0`.